Move utils method out of InstanceSubmitter
