+++ collect_app/src/test/java/org/odk/collect/android/utilities/InstanceUploaderUtilsTest.java
@@ -4,16 +4,18 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.odk.collect.formstest.InMemFormsRepository;
-import org.odk.collect.formstest.InMemInstancesRepository;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.instances.Instance;
+import org.odk.collect.formstest.FormUtils;
+import org.odk.collect.formstest.InMemFormsRepository;
+import org.odk.collect.formstest.InMemInstancesRepository;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.odk.collect.android.utilities.InstanceUploaderUtils.shouldFormBeSent;
 import static org.odk.collect.formstest.FormUtils.createXFormFile;
 
 @RunWith(AndroidJUnit4.class)
@@ -92,6 +94,64 @@ public class InstanceUploaderUtilsTest {
     public void doesUrlRefersToGoogleSheetsFileTest() {
         assertThat(InstanceUploaderUtils.doesUrlRefersToGoogleSheetsFile("https://docs.google.com/spreadsheets/d/169qibpJCWgUy-SRtoyvKd1EKwV1nDfM0/edit#gid=773120038"), is(true));
         assertThat(InstanceUploaderUtils.doesUrlRefersToGoogleSheetsFile("https://drive.google.com/file/d/169qibpJCWgUy-SRtoyvKd1EKwV1nDfM0/edit#gid=773120038"), is(false));
+    }
+
+    @Test
+    public void shouldFormBeSentFunction_shouldReturnFalseIfAutoSendNotSpecifiedOnFormLevelAndDisabledInSettings() {
+        InMemFormsRepository formsRepository = new InMemFormsRepository();
+
+        formsRepository.save(new Form.Builder()
+                .dbId(1L)
+                .formId("1")
+                .version("1")
+                .formFilePath(org.odk.collect.formstest.FormUtils.createXFormFile("1", "1").getAbsolutePath())
+                .build());
+
+        assertThat(shouldFormBeSent(formsRepository, "1", "1", false), is(false));
+    }
+
+    @Test
+    public void shouldFormBeSentFunction_shouldReturnTrueIfAutoSendNotSpecifiedOnFormLevelButEnabledInSettings() {
+        InMemFormsRepository formsRepository = new InMemFormsRepository();
+
+        formsRepository.save(new Form.Builder()
+                .dbId(1L)
+                .formId("1")
+                .version("1")
+                .formFilePath(org.odk.collect.formstest.FormUtils.createXFormFile("1", "1").getAbsolutePath())
+                .build());
+
+        assertThat(shouldFormBeSent(formsRepository, "1", "1", true), is(true));
+    }
+
+    @Test
+    public void shouldFormBeSentFunction_shouldReturnFalseIfAutoSendSpecifiedAsFalseOnFormLevelButEnabledInSettings() {
+        InMemFormsRepository formsRepository = new InMemFormsRepository();
+
+        formsRepository.save(new Form.Builder()
+                .dbId(1L)
+                .formId("1")
+                .version("1")
+                .autoSend("false")
+                .formFilePath(org.odk.collect.formstest.FormUtils.createXFormFile("1", "1").getAbsolutePath())
+                .build());
+
+        assertThat(shouldFormBeSent(formsRepository, "1", "1", true), is(false));
+    }
+
+    @Test
+    public void shouldFormBeSentFunction_shouldReturnTrueIfAutoSendSpecifiedAsTrueOnFormLevelButDisabledInSettings() {
+        InMemFormsRepository formsRepository = new InMemFormsRepository();
+
+        formsRepository.save(new Form.Builder()
+                .dbId(1L)
+                .formId("1")
+                .version("1")
+                .autoSend("true")
+                .formFilePath(FormUtils.createXFormFile("1", "1").getAbsolutePath())
+                .build());
+
+        assertThat(shouldFormBeSent(formsRepository, "1", "1", false), is(true));
     }
 
     private InMemInstancesRepository getTestInstancesRepository() {
